changes to the parameters of quart dimensions from scripts
We are choosing a range around 20 mm for quartz size z. We can look at the effect on detector resolution. And the width of quartz, quartz size x was estimated from the document : https://moller.jlab.org/DocDB/0007/000769/003/210817_MDtiling_CGal.pdf,slide 11, we got the radius of the rings,maximum and minimum and calculated the maximum and minimum circumference and then estimated the rings minimum and maximum width by dividing the circumference by the number of modules. If we choose the minimum width then there will be gaps in the design and if we take the maximum then we need to consider double counting. From the document : https://github.com/mtgericke/MOLLER-IntDetectorOpticalSim/blob/main/GREXMakeScanMacros.py we got the estimated value of quartz interface opening x and z and set it to (quartz size X + 8 and quartz size Z + 7).