Making adjustments to reflectivity reading to ensure that lower bounds on incident angles are taken instead of upper bounds
The existing implementation seems to take the higher value in the incident angle ranges 0 to 30, 30 to 45, 45 to 60, >60. For example, if the incident angle is 25 degree, it picks the reflectivity table for 30 degrees. The original implementation in Michael's code seems to take the lower value in the ranges. For an incident angle of 25 degrees, it would pick reflectivity table for 0 degrees. This change should mimic that original behavior [needs to be tested]